+++ install/migrations/0048_add_media_owned_to_downloads_table.sql
@@ -1,0 +1,1 @@
+ALTER TABLE {prefix}downloads ADD COLUMN media_owned TINYINT(1) NOT NULL DEFAULT 1 AFTER media_id;